Se agregó código para detectar errores en las entradas, y una opción para escoger el número de procesos en el reader

+++ fase2/parte1/conn/socket.h
@@ -13,7 +13,7 @@ class Socket{
 		int Read(char *, int);
 		int Write(const char *, int);
 		int Listen(int = SOMAXCONN);
-		int Bind(int);
+		int Bind(long);
 		Socket * Accept();
 		int Shutdown(int);
 		int Close();
